requirements.txt
opencv-python-headless
cvzone
pygame-ce